new architecture. add odometry, manipulator files. add launch file
